+++ adapters/src/exceptions/repository/__init__.py
@@ -1,0 +1,1 @@
+from .base import RepositoryException
