+++ api/__init__.py
@@ -3,11 +3,8 @@ from typing import AsyncGenerator
 
 from fastapi import FastAPI
 
-from adapters.src.repositories.sql.config_db import (SessionManager,
-                                                     SQLConnection)
-from api.src.routers.appointment import appointment_router
-from api.src.routers.index import index_router
-from api.src.routers.product import product_router
+from adapters.src.repositories.sql import SessionManager, SQLConnection
+from api.src.routers import appointment_router, index_router, product_router
 
 
 @asynccontextmanager
