+++ api/src/dtos/__init__.py
@@ -1,0 +1,2 @@
+from .appointment_dto import Appointment
+from .product_dto import Product
