+++ api/src/routers/__init__.py
@@ -1,0 +1,3 @@
+from .appointment import appointment_router
+from .index import index_router
+from .product import product_router
