+++ factories/use_cases/appointment.py
@@ -7,7 +7,7 @@ from factories.repositories.email_sender import email_sender_repository
 def create_appointment_use_case() -> CreateAppointment:
     return CreateAppointment(
         appointment_repository=sql_appointment_repository(),
-        email_sender_repository=email_sender_repository(),  # TODO: Use the real email repository when it's done
+        email_sender_repository=email_sender_repository(),
     )
